segmentation_baseline/requirements.txt
# Semantic segmentation baseline — RandLA-Net via Open3D-ML (PyTorch backend)
#
# Install PyTorch first from https://pytorch.org/get-started/locally/
# matching your CUDA version (or CPU-only), then install these:
#
#   pip install torch torchvision  # or with CUDA, e.g. +cu121
#   pip install -r requirements.txt

open3d>=0.18.0
open3d-ml                # RandLA-Net lives here; installs on top of open3d
laspy[laszip]>=2.5       # LAS/LAZ I/O
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3        # confusion matrix, metrics
tqdm>=4.65
matplotlib>=3.7          # optional, for confusion-matrix plots
